use view.py for roipoly in generate_mask.py
generate_mask:
+ add --keep-nan option to be able to not exclude pixels with nan values

move get_poly_mask() from utils/plot.py to utils/plot_ext.py to be able to call view.py functions for more advanced plotting capability to facilitate the polygon selection.
