+++ mintpy/utils/plot.py
@@ -180,108 +180,6 @@ class BasemapExt(Basemap):
 ######################################### BasemapExt class end ############################################
 
 
-################################# SelectFromCollection class begin ########################################
-class SelectFromCollection(object):
-    """Select indices from a matplotlib collection using `PolygonSelector`.
-    
-    Selected pixels within the polygon is marked as True and saved in the 
-    member variable self.mask, in the same size as input AxesImage object
-    with all the other pixels marked as False.
-
-    Parameters
-    ----------
-    ax : :class:`~matplotlib.axes.Axes`
-        Axes to interact with.
-
-    collection : :class:`matplotlib.collections.Collection` subclass
-        Collection you want to select from.
-
-    Examples
-    --------
-    import matplotlib.pyplot as plt
-    from mintpy.utils import readfile, plot as pp
-    
-    fig, ax = plt.subplots()
-    data = readfile.read('velocity.h5', datasetName='velocity')[0]
-    im = ax.imshow(data)
-    
-    selector = pp.SelectFromCollection(ax, im)
-    plt.show()
-    selector.disconnect()
-    
-    plt.figure()
-    plt.imshow(selector.mask)
-    plt.show()
-    """
-
-    def __init__(self, ax, collection, alpha_other=0.3):
-        from matplotlib.widgets import PolygonSelector
-        self.canvas = ax.figure.canvas
-        self.collection = collection
-        self.prepare_coordinates()
-
-        self.poly = PolygonSelector(ax, self.onselect)
-
-        msg = "\nSelect points in the figure by enclosing them within a polygon.\n"
-        msg += "Press the 'esc' key to start a new polygon.\n"
-        msg += "Try hold to left key to move a single vertex.\n"
-        msg += "After complete the selection, close the figure/window to continue.\n"
-        print(msg)
-
-    def prepare_coordinates(self):
-        imgExt = self.collection.get_extent()
-        self.length = int(imgExt[2] - imgExt[3])
-        self.width  = int(imgExt[1] - imgExt[0])
-        yy, xx = np.mgrid[:self.length, :self.width]
-        self.coords = np.hstack((xx.reshape(-1, 1),
-                                 yy.reshape(-1, 1)))
-
-    def onselect(self, verts):
-        from matplotlib.path import Path
-        self.poly_path = Path(verts)
-        self.mask = self.poly_path.contains_points(self.coords).reshape(self.length,
-                                                                        self.width)
-        self.canvas.draw_idle()
-
-    def disconnect(self):
-        self.poly.disconnect_events()
-        self.canvas.draw_idle()
-
-## Utility script
-def get_poly_mask(data, print_msg=True):
-    """Get mask of pixels within polygon from interactive selection
-    Parameters: data : 2D np.array in size of (length, width)
-    Returns:    mask : 2D np.arrat in size of (length, width) in np.bool_ format
-    """
-    # disp min/max
-    vlim = np.nanmax(np.abs(data))
-    vmin, vmax = -vlim, vlim
-    # for dataset with non-negative values such as elevation
-    if np.nanmin(data) > 0:
-        vmin = np.nanmin(data)
-
-    # plot
-    fig, ax = plt.subplots()
-    im = ax.imshow(data, cmap='jet', vmin=vmin, vmax=vmax)
-    fig.colorbar(im)
-
-    selector = SelectFromCollection(ax, im)
-    plt.show()
-    selector.disconnect()
-
-    if hasattr(selector, 'mask'):
-        mask = selector.mask
-        if print_msg:
-            print('selected polygon: {}'.format(selector.poly_path))
-    else:
-        mask = None
-        if print_msg:
-            print('no polygon selected.\n')
-    return mask
-################################## SelectFromCollection class end #########################################
-
-
-
 ########################################### Parser utilities ##############################################
 def cmd_line_parse(iargs=''):
     parser = argparse.ArgumentParser(description='Ploting Parser')
